Move .NET 4.5 unit tests into UnitTest project, and remove UnitTest45 project
