--- Hyprlinkr.UnitTest45/packages.config
@@ -1,12 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<packages>
-  <package id="AutoFixture" version="3.16.4" targetFramework="net45" />
-  <package id="AutoFixture.AutoMoq" version="3.16.4" targetFramework="net45" />
-  <package id="AutoFixture.Xunit" version="3.16.4" targetFramework="net45" />
-  <package id="Microsoft.AspNet.WebApi.Client" version="5.2.3" targetFramework="net45" />
-  <package id="Microsoft.AspNet.WebApi.Core" version="5.2.3" targetFramework="net45" />
-  <package id="Moq" version="4.2.1312.1622" targetFramework="net45" />
-  <package id="Newtonsoft.Json" version="6.0.4" targetFramework="net45" />
-  <package id="xunit" version="1.9.2" targetFramework="net45" />
-  <package id="xunit.extensions" version="1.9.2" targetFramework="net45" />
-</packages>
