+++ Hyprlinkr.sln
@@ -1,7 +1,7 @@
 ﻿
 Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio 2013
-VisualStudioVersion = 12.0.30110.0
+VisualStudioVersion = 12.0.31101.0
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = ".nuget", ".nuget", "{7A13823C-0B72-4B86-9E75-76B44B9BB339}"
 	ProjectSection(SolutionItems) = preProject
@@ -15,8 +15,6 @@ EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ExampleService", "ExampleService\ExampleService.csproj", "{E1B8892B-122A-4E8B-9C45-0602D7A83048}"
 EndProject
 Project("{F2A71F9B-5D33-465A-A702-920D77279786}") = "Hyprlinkr.UnitTest.FSharp", "Hyprlinkr.UnitTest.FSharp\Hyprlinkr.UnitTest.FSharp.fsproj", "{DA4CC787-3FAF-4929-B372-BDC118E2E8E6}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Hyprlinkr.UnitTest45", "Hyprlinkr.UnitTest45\Hyprlinkr.UnitTest45.csproj", "{C618B2E6-96D8-49F2-8DFC-2FB9DFAEDDAD}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -49,12 +47,6 @@ Global
 		{DA4CC787-3FAF-4929-B372-BDC118E2E8E6}.Release|Any CPU.Build.0 = Release|Any CPU
 		{DA4CC787-3FAF-4929-B372-BDC118E2E8E6}.Verify|Any CPU.ActiveCfg = Release|Any CPU
 		{DA4CC787-3FAF-4929-B372-BDC118E2E8E6}.Verify|Any CPU.Build.0 = Release|Any CPU
-		{C618B2E6-96D8-49F2-8DFC-2FB9DFAEDDAD}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{C618B2E6-96D8-49F2-8DFC-2FB9DFAEDDAD}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{C618B2E6-96D8-49F2-8DFC-2FB9DFAEDDAD}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{C618B2E6-96D8-49F2-8DFC-2FB9DFAEDDAD}.Release|Any CPU.Build.0 = Release|Any CPU
-		{C618B2E6-96D8-49F2-8DFC-2FB9DFAEDDAD}.Verify|Any CPU.ActiveCfg = Release|Any CPU
-		{C618B2E6-96D8-49F2-8DFC-2FB9DFAEDDAD}.Verify|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
